clustermesh-apiserver: include kvstoremesh as subcommand
Merge the kvstoremesh logic into the clustermesh-apiserver as a separate
subcommand, so that we can get rid of one image and reduce the overall
time to bootstrap the clustermesh-apiserver deployment. The additional
boilerplate to build the kvstoremesh images is dropped.

Size of the images before this change:
$ make docker-clustermesh-apiserver-image
$ docker inspect -f '{{ .Size }}' quay.io/cilium/clustermesh-apiserver:latest | \
         numfmt --to=si --suffix=B --format="%.2f"
---> 72.09MB
$ make docker-kvstoremesh-image
$ docker inspect -f '{{ .Size }}' quay.io/cilium/kvstoremesh:latest | \
         numfmt --to=si --suffix=B --format="%.2f"
---> 64.62MB

Size of the image after this change:
$ make docker-clustermesh-apiserver-image
$ docker inspect -f '{{ .Size }}' quay.io/cilium/clustermesh-apiserver:latest | \
         numfmt --to=si --suffix=B --format="%.2f"
---> 72.26MB

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>