service: make global services validation backward compatible
2819d7c4cc27 ("service: add extra validation for global services")
introduced a validation function to prevent unmarshalling invalid global
service objects retrieved from remote clusters. In particular, it
flags a ClusterID=0 as invalid, unless it matches the one of the local
cluster (as that case is supported for external workloads).

Yet, the ClusterID field was not present as part of the ClusterService
struct in v1.13, causing a failure when the v1.14 agents connect to
a cluster still running the v1.13 clustermesh-apiserver. Hence, let's
relax this validation to support ClusterID=0 and prevent issues during
upgrades.

Fixes: 2819d7c4cc27 ("service: add extra validation for global services")
Reported-by:  Luke <lifeixiangchina@gmail.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>